+++ BookMyShow.BuinessLogicLayer/Managers/TheaterManager.cs
@@ -3,9 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using BookMyShow.BuinessLogicLayer.CustomExceptions;
 using BookMyShow.BuinessLogicLayer.DTOs;
 using BookMyShow.DataAccessLayer.Abstract;
 using BookMyShow.DataAccessLayer.Models;
+using Microsoft.Identity.Client;
 
 namespace BookMyShow.BuinessLogicLayer.Managers
 {
@@ -26,40 +28,60 @@ namespace BookMyShow.BuinessLogicLayer.Managers
             var result = await _theaterService.GetTheaterById(id);
             if (result == null)
             {
-                return null;
+                throw new Exception("Theater does not exist for the provided id");
             }
             return TheaterDto.MapToDto(result);
         }
 
         public async Task AddTheater(TheaterDto theaterDto)
         {
-            if (theaterDto != null)
+            var exceptions = new List<string>();
+            var inputTheaterName = theaterDto.TheaterName.Trim();
+
+            if(inputTheaterName.Length <= 8) { exceptions.Add("Theater name should be more than or equals to 8 characters"); }
+
+            var isExist = await _theaterService.GetTheaterByName(inputTheaterName);
+            if(isExist != null) { exceptions.Add("Theater name already exist"); }
+
+            if (exceptions.Count > 0) throw new CustomException(exceptions);
+
+            var theater = new Theater
             {
-                var theater = new Theater
-                {
-                    TheaterName = theaterDto.TheaterName,
-                    IsMultiScreen = theaterDto.IsMultiScreen,
-                    CreatedBy = 1
-                };
-                await _theaterService.AddTheater(theater);
-            }
+                TheaterName = inputTheaterName,
+                IsMultiScreen = theaterDto.IsMultiScreen,
+                CreatedBy = 1
+            };
+
+            await _theaterService.AddTheater(theater);
         }
 
         public async Task UpdateTheater(int id, TheaterDto theaterDto)
         {
-            if (theaterDto != null)
-            {
-                var theater = await _theaterService.GetTheaterById(id);
-                theater.TheaterName = theaterDto.TheaterName;
-                theater.IsMultiScreen = theaterDto.IsMultiScreen;
-                theater.ChangedBy = 1;
-                theater.ChangedOn = DateTime.Now;
-                await _theaterService.UpdateTheater(theater);
-            }
+            var exceptions = new List<string>();
+            var inputTheaterName = theaterDto.TheaterName.Trim();
+
+            if (inputTheaterName.Length <= 8) { exceptions.Add("Theater name should be more than or equals to 8 characters"); }
+
+            var isExist = await _theaterService.GetTheaterByName(inputTheaterName);
+            if (isExist != null) { exceptions.Add("Theater name already exist"); }
+
+            var theater = await _theaterService.GetTheaterById(id);
+            if (theater == null) { exceptions.Add("Theater does not exist for the provided id"); }
+
+            if (exceptions.Count > 0) throw new CustomException(exceptions);
+
+            theater.TheaterName = theaterDto.TheaterName;
+            theater.IsMultiScreen = theaterDto.IsMultiScreen;
+            theater.ChangedBy = 1;
+            theater.ChangedOn = DateTime.Now;
+
+            await _theaterService.UpdateTheater();
         }
 
         public async Task DeleteTheater(int id)
         {
+            var theater = await _theaterService.GetTheaterById(id);
+            if (theater == null) throw new Exception("Theater does not exist for the provided id");
             await _theaterService.DeleteTheater(id);
         }
